+++ pykblib/const.py
@@ -1,3 +1,3 @@
 """PyKBLib constants."""
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"