+++ CHANGELOG.rst
@@ -3,7 +3,21 @@ Release History
 
 In-Development
 --------------
-- Refactor Team class to use the Keybase Team API.
+- Expand Team functionality.
+
+0.2.1 (2019.04.01)
+------------------
+- Implemented functions for communicating with the various APIs.
+- Updated the _get_memberships function to allow for a specified username.
+- Added Team.add_members function.
+- Updated Team.add_member function to use Team.add_members function.
+- Updated documentation to reflect the addition of Team.add_members.
+- Updated Team.change_member_role to use the Team API.
+- Updated Team.remove_member to use the Team API.
+- Updated Team class to take parent Keybase instance as an argument. This will speed up Team creation slightly.
+- Updated Team.update function to use the Team API.
+- Updated Team documentation to reflect that the Team.members list will not include deleted accounts.
+- Refactored Team class to use the Keybase Team API.
 
 0.2.0 (2019.04.01)
 ------------------